Update imports in `ComposeNewsTvApplication.kt` and `MainViewModel.kt` to reflect the move of `dispatcherModule` and `DispatcherProvider` to `core:base`
